% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_handlers.R
\name{DBI_creds_edit_handler}
\alias{DBI_creds_edit_handler}
\title{DBI edit to the database handler}
\usage{
DBI_creds_edit_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which need to contain within its data:
\itemize{
\item username
\item password
}

It can also contain elements for change:
\itemize{
\item new_username
\item new_email
\item new_password
}}
}
\description{
DBI edit to the database handler
}
\seealso{
Other DBI handler functions: 
\code{\link{DBI_login_handler}()},
\code{\link{DBI_register_handler}()}
}
\concept{DBI handler functions}
