% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_handlers.R
\name{DBI_login_handler}
\alias{DBI_login_handler}
\title{DBI login handler}
\usage{
DBI_login_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item password
}}
}
\description{
DBI login handler
}
\seealso{
Other DBI handler functions: 
\code{\link{DBI_creds_edit_handler}()},
\code{\link{DBI_register_handler}()}
}
\concept{DBI handler functions}
