% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_handlers.R
\name{DBI_register_handler}
\alias{DBI_register_handler}
\title{DBI register handler}
\usage{
DBI_register_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item password
\item email
}}
}
\description{
Default handler function querying database to confirm registration
validity and input new data. Used within object of \code{RegLogDBIConnector} class internally.
}
\seealso{
Other DBI handler functions: 
\code{\link{DBI_credsEdit_handler}()},
\code{\link{DBI_login_handler}()},
\code{\link{DBI_resetPass_confirmation_handler}()},
\code{\link{DBI_resetPass_generation_handler}()}
}
\concept{DBI handler functions}
\concept{DBI_handler}
\keyword{internal}
