% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_handlers.R
\name{DBI_resetPass_generation_handler}
\alias{DBI_resetPass_generation_handler}
\title{DBI resetpass code generation handler}
\usage{
DBI_resetPass_generation_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which need to contain within its data:
\itemize{
\item username
}}
}
\description{
Default handler function querying database to confirm credentials
edit procedure and update values saved within database. Used within object of
\code{RegLogDBIConnector} class internally.
}
\seealso{
Other DBI handler functions: 
\code{\link{DBI_credsEdit_handler}()},
\code{\link{DBI_login_handler}()},
\code{\link{DBI_register_handler}()},
\code{\link{DBI_resetPass_confirmation_handler}()}
}
\concept{DBI handler functions}
\concept{DBI_handler}
\keyword{internal}
