% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_texts.R
\name{reglog_txt}
\alias{reglog_txt}
\title{Get texts for given language. Used within deprecated \code{login_server}}
\usage{
reglog_txt(lang, x = NULL, custom_txts = NULL)
}
\arguments{
\item{lang}{character to identify the language}

\item{x}{character to identify the txt to get. If NULL, all labels are
recovered}

\item{custom_txts}{named list providing custom messages to replace default for
specific languages.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
'RegLog_txt' outside of internal usage should be used only for getting the
structure of all texts generated by 'shiny.reglog'.

To customize texts used by RegLog objects, provide within their
call named list to the 'custom_txts' argument - it will be passed to 'custom_txts'
within this call. You can check validity of your list by providing the 'custom_txts'
and calling this function in console.

Values of list provided should be named in the same way as the default text
you are willing to replace.
}
\keyword{internal}
