% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_handlers.R
\name{emayili_mail_handler}
\alias{emayili_mail_handler}
\title{Emayili email sending handler}
\usage{
emayili_mail_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item email
\item app_name
\item app_address
\item reset_code (optional for reset code mails)
}

It can also contain \emph{mail_subject} and \emph{mail_body} if you want to send custom
\code{RegLogEmayiliConnector} message.}
}
\description{
Default handler function parsing and sending email.
Used within object of \code{RegLogEmayiliConnector} class internally.
}
\seealso{
Other mail handler functions: 
\code{\link{gmailr_mail_handler}()}
}
\concept{mail handler functions}
\keyword{internal}
