% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_handlers.R
\name{emayili_register_handler}
\alias{emayili_register_handler}
\title{Emayili register handler}
\usage{
emayili_register_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item email
}}
}
\description{
Default handler function parsing and sending register confirmation
email to newly registered user of the package. Used within object of
\code{RegLogEmayiliConnector} class internally.
}
\concept{emayili handler functions}
