% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_handlers.R
\name{emayili_resetPass_handler}
\alias{emayili_resetPass_handler}
\title{Emayili password reset email handler}
\usage{
emayili_resetPass_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item email
\item reset_code
\item app_name
\item app_address
}

It can also contain \emph{mail_subject} and \emph{mail_body} if you want to send custom
\code{RegLogEmayiliConnector} message.}
}
\description{
Default handler function parsing and sending email confirming registration.
Used within object of \code{RegLogEmayiliConnector} class internally.
}
\seealso{
Other mail handler functions: 
\code{\link{emayili_register_handler}()},
\code{\link{gmailr_send_handler}()}
}
\concept{mail handler functions}
