% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_handlers.R
\name{gmailr_custom_mail_handler}
\alias{gmailr_custom_mail_handler}
\title{Gmailr custom email sending handler}
\usage{
gmailr_custom_mail_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item process
\item username
\item email
\item mail_subject
\item mail_body
\item mail_attachment (optional)
}}
}
\value{
\code{RegLogEmayiliConnector} message.
}
\description{
Default handler function parsing and sending email.
Used within object of \code{RegLogGmailrConnector} class internally. It can
send custom emails using subject, body and attachments from
\code{RegLogConnectorMessage} of type \code{custom_mail}.
}
\seealso{
Other mail handler functions: 
\code{\link{emayili_custom_mail_handler}()},
\code{\link{emayili_reglog_mail_handler}()},
\code{\link{gmailr_reglog_mail_handler}()}
}
\concept{mail handler functions}
\concept{mail_handler}
\keyword{internal}
