% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_handlers.R
\name{gmailr_send_handler}
\alias{gmailr_send_handler}
\title{Gmailr send email handler}
\usage{
gmailr_send_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item email
\item mail_subject
\item mail_body
}}
}
\description{
Default handler function parsing and sending register confirmation
email to newly registered user of the package. Used within object of
\code{RegLogEmayiliConnector} class internally.
}
\seealso{
Other mail handler functions: 
\code{\link{emayili_register_handler}()},
\code{\link{emayili_resetPass_handler}()}
}
\concept{mail handler functions}
