% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet_handlers.R
\name{gsheet_credsEdit_handler}
\alias{gsheet_credsEdit_handler}
\title{googlesheets edit to the database handler}
\usage{
gsheet_credsEdit_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which need to contain within its data:
\itemize{
\item account_id
\item password
}

It can also contain elements for change:
\itemize{
\item new_username
\item new_email
\item new_password
}}
}
\description{
Default handler function querying database to confirm credentials
edit procedure and update values saved within database. Used within object of
\code{RegLogGsheetConnector} class internally.
}
\seealso{
Other googlesheets handler functions: 
\code{\link{gsheet_login_handler}()},
\code{\link{gsheet_register_handler}()},
\code{\link{gsheet_resetPass_confirmation_handler}()},
\code{\link{gsheet_resetPass_generation_handler}()}
}
\concept{googlesheets handler functions}
\concept{gsheet_handler}
\keyword{internal}
