% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsheet_handlers.R
\name{gsheet_login_handler}
\alias{gsheet_login_handler}
\title{googlesheets login handler}
\usage{
gsheet_login_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item password
}}
}
\description{
Default handler function querying database to confirm login
procedure. Used within object of \code{RegLogGsheetConnector} class internally.
}
\seealso{
Other googlesheets handler functions: 
\code{\link{gsheet_credsEdit_handler}()},
\code{\link{gsheet_register_handler}()},
\code{\link{gsheet_resetPass_confirmation_handler}()},
\code{\link{gsheet_resetPass_generation_handler}()}
}
\concept{googlesheets handler functions}
\concept{gsheet_handler}
\keyword{internal}
