% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_handlers.R
\name{mailMessageAttachment}
\alias{mailMessageAttachment}
\title{Mail attachment data to be handled by mailConnector via \code{custom_mail} RegLogConnectorMessage}
\usage{
mailMessageAttachment(filepath, filename = NULL, cid = NULL, filetype = NULL)
}
\arguments{
\item{filepath}{path to the file to be attached}

\item{filename}{name of the file to be used (supported by RegLogEmayiliConnector)}

\item{cid}{content ID to be used to access in email body}

\item{filetype}{mime type of the attached file}
}
\value{
mailMessageAttachment object
}
\description{
Mail attachment data to be handled by mailConnector via \code{custom_mail} RegLogConnectorMessage
}
