% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_UI.R
\name{register_UI}
\alias{register_UI}
\title{Shiny UI module for registration box}
\usage{
register_UI(id = "login_system", lang = "en")
}
\arguments{
\item{id}{the id of the module. Defaults to "login_system" for all of the modules contained within the package. If you plan to use serveral login systems inside your app or for any other reason need to change it, remember to keep consistent id for all elements of module.}

\item{lang}{specifies the app used language. Defaults to "en" for English. Package also supports "pl" for Polish}
}
\value{
NONE
}
\description{
This function creates a UI div() element containing informations and input necessary for user to register new account.
As it outputs a div() element, you can put it inside container of your choosing (be it some tabItem, fluidPage, fluidRow etc.)
}
\details{
It need to be used in conjuction with \code{login_server()} function and is suggested to be used alongside \code{login_UI()} and \code{password_reset_UI()} for full potential.
}
\seealso{
login_server() for more details and example
}
