% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_create.R
\name{RegLog_DBI_database_create}
\alias{RegLog_DBI_database_create}
\title{Create RegLog-valid database tables with DBI}
\usage{
RegLog_DBI_database_create(
  conn,
  user_name = "user",
  reset_code_name = "reset_code",
  use_log = FALSE,
  log_name = "logs"
)
}
\arguments{
\item{conn}{DBI connection object}

\item{user_name}{Name of table for storing user credentials. Defaults to
'user'. Mandatory table.}

\item{reset_code_name}{Name of the table for storing generated password reset
codes. Defaults to 'reset_code'. Mandatory table.}

\item{use_log}{Should the table for keeping RegLogServer logs be
also created? Defaults to FALSE}

\item{log_name}{Name of the table for storing logs from RegLogServer object.
Used only if \code{use_log = TRUE}. Defaults to \code{logs}}
}
\description{
Create RegLog-valid database tables with DBI
}
\details{
Currently, the function is tested and working correctly for
SQLite and MySQL databases. If you want to use another DBI-supported
database, you need to create tables in other ways.

Created tables should have following structure:
\itemize{
\item user (default name)
\itemize{
\item id: integer, primary key, auto-increment
\item username: varchar(255), NOT NULL, unique
\item password: varchar(255), NOT NULL
\item email: varchar(255), NOT NULL, unique
\item create_time: datetime, NOT NULL
\item update_time: datetime, NOT NULL
}
\item reset_code (default name)
\itemize{
\item id: integer, primary key, auto-increment
\item user_id: integer, NOT NULL
\item reset_code: varchar(10), NOT NULL
\item used: tinyint, NOT NULL
\item create_time: datetime, NOT NULL
\item update_time: datetime, NOT NULL
}
\item logs (default name, optional)
\itemize{
\item id: integer, primary key, auto-increment
\item time: numeric, size 15, 5 decimal, NOT NULL
\item session: varchar(255), NOT NULL
\item direction: varchar(255), NOT NULL
\item type: varchar(255), NOT NULL
\item note: varchar(255)
}
}
}
\concept{RegLog databases}
