% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLogUI.R
\name{RegLog_register_UI}
\alias{RegLog_register_UI}
\title{Generate Register UI for RegLog system}
\usage{
RegLog_register_UI(module_id = "login_system")
}
\arguments{
\item{module_id}{Character declaring the id of the module. Defaults to 'login_system'.
Recommended to keep it that way, unless it would cause any namespace issues.}
}
\description{
Generate Register UI for RegLog system
}
\seealso{
Other RegLog UI: 
\code{\link{RegLog_edit_UI}()},
\code{\link{RegLog_login_UI}()},
\code{\link{RegLog_resetPass_confirm_code_UI}()},
\code{\link{RegLog_resetPass_send_code_UI}()}
}
\concept{RegLog UI}
