% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLogUI.R
\name{RegLog_resetPass_send_code_UI}
\alias{RegLog_resetPass_send_code_UI}
\title{Generate ResetPass code send UI for RegLog system}
\usage{
RegLog_resetPass_send_code_UI(module_id = "login_system")
}
\arguments{
\item{module_id}{Character declaring the id of the module. Defaults to 'login_system'.
Recommended to keep it that way, unless it would cause any namespace issues.}
}
\description{
Generate ResetPass code send UI for RegLog system
}
\details{
Password reset procedure needs two-step UI: firstly, provide username
and email provided during registration. Secondly, provide reset code send via
email and new password. This function binds the UI for the first step.
}
\seealso{
Other RegLog UI: 
\code{\link{RegLog_edit_UI}()},
\code{\link{RegLog_login_UI}()},
\code{\link{RegLog_register_UI}()},
\code{\link{RegLog_resetPass_confirm_code_UI}()}
}
\concept{RegLog UI}
