% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegLogUI.R
\name{RegLog_resetPass_UI}
\alias{RegLog_resetPass_UI}
\title{Generate ResetPass code UI for RegLog system}
\usage{
RegLog_resetPass_UI(module_id = "login_system")
}
\arguments{
\item{module_id}{Character declaring the id of the module. Defaults to 'login_system'.
Recommended to keep it that way, unless it would cause any namespace issues.}
}
\description{
Generate ResetPass code UI for RegLog system
}
\seealso{
Other RegLog UI: 
\code{\link{RegLog_credsEdit_UI}()},
\code{\link{RegLog_login_UI}()},
\code{\link{RegLog_register_UI}()}
}
\concept{RegLog UI}
