% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texts.R
\name{reglog_txt}
\alias{reglog_txt}
\title{Get texts for given language and supports getting them back}
\usage{
reglog_txt(lang, x = NULL, custom_txts = NULL)
}
\arguments{
\item{lang}{character to identify the language}

\item{x}{character to identify the txt to get. If NULL, all labels are
recovered}

\item{custom_txts}{named list providing custom messages to replace default for
specific languages.}
}
\description{
Get texts for given language and supports getting them back
}
\details{
'quetzio_txt' outside of internal usage should be used only for getting the
structure of all texts generated by 'shiny.reglog'. You then call it in
your console without specifying the 'private' argument.

To customize texts used by reglog modules, provide within their
call named list to the 'custom_txts' argument - it will be passed to 'custom_txts'
within this call. You can check validity of your list by providing the 'custom_txts'
and calling this function in console.

Values of list provided should be named in the same way as the default text
you are willing to replace.
}
