% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{RegLogDemo}
\alias{RegLogDemo}
\title{Demonstration ShinyApp with basic RegLog system}
\usage{
RegLogDemo(emayili_smtp = NULL, emayili_from = NULL)
}
\arguments{
\item{emayili_smtp}{defined emayili smtp server for you e-mail provider.
If kept as default NULL, the e-mail sending functionality won't
be used. If provided, it will require an installation of 'emayili' package.}

\item{emayili_from}{String containing e-mail from which thesending will take place.
Used only with 'emayili_smtp' defined.}
}
\description{
You can play a little with RegLogSever functionalities launching this ShinyApp.
This demo needs also an installation of 'RSQLite' package to create and
menage a temporary database.
}
