% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.reglog-package.R
\docType{package}
\name{shiny.reglog-package}
\alias{shiny.reglog}
\alias{shiny.reglog-package}
\title{shiny.reglog: Optional Login and Registration Module System for ShinyApps}
\description{
RegLog system provides a set of shiny modules to handle register procedure for your users, alongside with login, edit credentials and password reset functionality. It provides support for popular SQL databases and optionally googlesheet-based database for easy setup. For email sending it provides support for 'emayili' and 'gmailr' backends. Architecture makes customizing usability pretty straightforward. The authentication system created with shiny.reglog is designed to be optional: user don't need to be logged-in to access your application, but when logged-in the user data can be used to read from and write to relational databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://statismike.github.io/shiny.reglog/}
}

}
\author{
\strong{Maintainer}: Michal Kosinski \email{kosinski.mich@gmail.com} (\href{https://orcid.org/0000-0002-8426-3654}{ORCID})

}
\keyword{internal}
