% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{sqlite_get_db}
\alias{sqlite_get_db}
\title{Function to read SQLite shiny.reglog database}
\usage{
sqlite_get_db(sqlite_db)
}
\arguments{
\item{sqlite_db}{path to your 'SQLite' database}
}
\value{
list containing \code{user_db} and \code{reset_db} dataframes
}
\description{
Function to read SQLite shiny.reglog database
}
\seealso{
create_sqlite_db
}
