% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBI_handlers.R
\name{DBI_login_handler}
\alias{DBI_login_handler}
\title{DBI login handler}
\usage{
DBI_login_handler(self, private, message)
}
\arguments{
\item{self}{R6 object element}

\item{private}{R6 object element}

\item{message}{RegLogConnectorMessage which should contain within its data:
\itemize{
\item username
\item password
}}
}
\description{
Default handler function querying database to confirm login
procedure. Used within object of \code{RegLogDBIConnector} class internally.
}
\seealso{
Other DBI handler functions: 
\code{\link{DBI_credsEdit_handler}()},
\code{\link{DBI_register_handler}()},
\code{\link{DBI_resetPass_confirmation_handler}()},
\code{\link{DBI_resetPass_generation_handler}()}
}
\concept{DBI handler functions}
\concept{DBI_handler}
\keyword{internal}
