% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{gsheet_get_db}
\alias{gsheet_get_db}
\title{Function to read googlesheets shiny.reglog database}
\usage{
gsheet_get_db(gsheet_db)
}
\arguments{
\item{gsheet_db}{ID of your 'googlesheets' database}
}
\value{
list containing \code{user_db} and \code{reset_db} dataframes
}
\description{
Function to read googlesheets shiny.reglog database
}
\seealso{
create_sqlite_db
}
