% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_UI.R
\name{logout_button}
\alias{logout_button}
\title{Action Button for logging out}
\usage{
logout_button(id = "login_system", lang = "en")
}
\arguments{
\item{id}{the id of the module. Defaults to "login_system" for all of the modules contained within the package. If you plan to use serveral login systems inside your app or for any other reason need to change it, remember to keep consistent id for all elements of module.}

\item{lang}{lang specifies the app used language. Accepts "en" or "pl". Defaults to "en"}
}
\value{
'actionButton' for logout
}
\description{
Simple action button - after pressing it, the user will be logged out
}
