% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.reglog-package.R
\docType{package}
\name{shiny.reglog-package}
\alias{shiny.reglog}
\alias{shiny.reglog-package}
\title{shiny.reglog: Optional Login and Registration Module System for ShinyApps}
\description{
Package creates UI modules for Login, Register and Password Reset boxes and server module for reading and writing user database contained within googlesheet or sqlite file. Creates automatic e-mail messages after registration and for resetting password procedure. Currently supports English language (default) and Polish. The authentication system created with shiny.reglog is designed to be optional: user don't need to be logged-in to access your application, but when logged-in the user data can be used to read from and write to relational databases.
}
\author{
\strong{Maintainer}: Michal Kosinski \email{kosinski.mich@gmail.com} (\href{https://orcid.org/0000-0002-8426-3654}{ORCID})

}
\keyword{internal}
